requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
comtypes==1.4.13
dataclasses-json==0.6.7
exceptiongroup==1.3.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
keyboard==0.13.5
langchain==0.3.12
langchain-classic==1.0.0
langchain-community==0.3.12
langchain-core==0.3.25
langchain-ollama==0.2.1
langchain-text-splitters==0.3.3
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.2.11
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
MouseInfo==0.1.3
mss==10.1.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
ollama==0.6.1
opencv-python==4.12.0.88
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pillow==12.0.0
propcache==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
SpeechRecognition==3.14.4
SQLAlchemy==2.0.45
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
